Expose the SQLContext and DataFrame to make it easier to use from PySpark
